warning fixes for launcher directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23640 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/launcher/rpc/LauncherState.C
@@ -19,7 +19,7 @@ LauncherState::SetupComponentRPCs(Xfer* xfer)
     //
     // Connect the RPCs to the xfer object.
     //
-    for(int i = 0; i < GetNumStateObjects(); ++i)
+    for(size_t i = 0; i < GetNumStateObjects(); ++i)
         xfer->Add(GetStateObject(i));
 }
 
@@ -32,7 +32,7 @@ LauncherState::GetNumStateObjects()
 VisItRPC*
 LauncherState::GetStateObject(int i)
 {
-    return (i >= 0 && i < objVector.size()) ?
+    return (i >= 0 && (size_t)i < objVector.size()) ?
            objVector[i] : 0;
 }
